fix: one read per source, and receipts that name the stating front end
Two Codex findings, plus the same defect wherever else it appeared:

- `SuppressionSource.from_file` digested the file with its own
  `read_bytes()` and then had `SuppressionList.load` read it again, so a
  file changed in between paired one content's digest with another's rules.
  `SuppressionList.load` now captures `source_sha256` during its single
  read (over raw bytes, so CRLF and LF twins differ), and the source takes
  the digest from the same load that produced the rules.
- An API-stated `scope_public` recorded `--scope-public-headers` in its
  receipt. `cross_front_end_differences` normalizes option spellings on
  purpose, so it cannot catch this — the receipt has to be right on its
  own. One `spell()` helper now decides the selector spelling per front
  end, applied to every explicit entry (`--contract`/`contract_mode`,
  `--policy`/`policy`, `--policy-file`/`policy_file_path`,
  `--public-symbol`/`force_public_symbols`, `--suppress`/`suppress`,
  `--pack`/`pack_paths`), not just the two already fixed.

Split the front-end test module once it passed the AI-readiness soft size
limit: resolution-and-precedence stays in
`test_compatibility_evaluation_frontend.py`, receipt/replay coverage moves
to `test_compatibility_evaluation_receipts.py`, with shared fixtures in
`_compat_eval_fixtures.py` (the `_detector_mutations.py` pattern).
