+++ tests/test_compatibility_evaluation_frontend.py
@@ -24,11 +24,15 @@ fields, and Phase 1's own gate: a CLI run and the equivalent typed
 
 from __future__ import annotations
 
-import hashlib
 import textwrap
 from pathlib import Path
 
 import pytest
+from _compat_eval_fixtures import (
+    policy_file as _policy_file,
+    resolve as _resolve,
+    write_pack as _write_pack,
+)
 
 from abicheck.api_types import CompareRequest, InputSpec
 from abicheck.buildsource.inline import load_build_config
@@ -55,7 +59,6 @@ from abicheck.compatibility_evaluation_frontend import (
     compatibility_config_from_compare_request,
     cross_front_end_differences,
     cross_front_end_equivalent,
-    resolve_compatibility_evaluation_config,
     severity_preset_identity,
 )
 from abicheck.compatibility_evaluation_packs import PackKind
@@ -66,31 +69,7 @@ from abicheck.compatibility_evaluation_resolver import (
 from abicheck.compatibility_evaluation_wiring import resolve_pack_field_assignments
 from abicheck.contract_relevance_types import ContractMode, SelectorLayer
 from abicheck.errors import PackManifestError
-from abicheck.policy_file import PolicyFile
 from abicheck.severity import SeverityLevel, resolve_severity_config
-
-
-def _write_pack(path: Path, *, pack_id: str, kind: str, assignments: str) -> Path:
-    path.write_text(
-        textwrap.dedent(f"""\
-            id: {pack_id}
-            version: 1
-            kind: {kind}
-            assignments:
-{textwrap.indent(assignments.strip(), "              ")}
-            """)
-    )
-    return path
-
-
-def _policy_file(tmp_path: Path, body: str) -> PolicyFile:
-    path = tmp_path / "policy.yml"
-    path.write_text(textwrap.dedent(body))
-    return PolicyFile.load(path)
-
-
-def _resolve(**kwargs):
-    return resolve_compatibility_evaluation_config(**kwargs)
 
 
 class TestDefaults:
@@ -969,391 +948,6 @@ class TestPackSuppliedSeverityActivatesAuto:
         assert cfg.gate.exit_code_scheme == "severity"
 
 
-class TestReplayProvenance:
-    """ADR-049 D6: a file-derived receipt entry identifies the exact content."""
-
-    def test_policy_file_derived_entries_carry_the_file_digest(self, tmp_path):
-        pf = _policy_file(
-            tmp_path,
-            """\
-            base_policy: sdk_vendor
-            internal_namespaces:
-              - detail
-            overrides:
-              soname_changed: warn
-            """,
-        )
-        expected = pf.source_sha256
-        cfg = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=pf))
-        assert expected is not None and len(expected) == 64
-        for field_name in (
-            POLICY_BASE_FIELD,
-            POLICY_OVERRIDES_FIELD,
-            INTERNAL_NAMESPACES_FIELD,
-        ):
-            assert cfg.provenance[field_name].sha256 == expected, field_name
-
-    def test_the_digest_identifies_what_was_parsed_not_what_is_on_disk_now(
-        self, tmp_path
-    ):
-        # The digest is captured by PolicyFile.load over the bytes it parsed.
-        # Editing the file afterwards must NOT change what this run's receipt
-        # records: re-reading would name content the run never applied.
-        pf = _policy_file(tmp_path, "base_policy: sdk_vendor\n")
-        before = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=pf))
-        (tmp_path / "policy.yml").write_text("base_policy: plugin_abi\n")
-        after = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=pf))
-        assert (
-            before.provenance[POLICY_BASE_FIELD].sha256
-            == after.provenance[POLICY_BASE_FIELD].sha256
-        )
-        assert after.policy.base.id == "sdk_vendor"
-
-    def test_reloading_an_edited_policy_file_does_change_the_digest(self, tmp_path):
-        pf = _policy_file(tmp_path, "base_policy: sdk_vendor\n")
-        before = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=pf))
-        reloaded = _policy_file(tmp_path, "base_policy: plugin_abi\n")
-        after = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=reloaded))
-        assert (
-            before.provenance[POLICY_BASE_FIELD].sha256
-            != after.provenance[POLICY_BASE_FIELD].sha256
-        )
-
-    def test_a_caller_supplied_digest_is_used_verbatim(self, tmp_path):
-        pf = _policy_file(tmp_path, "base_policy: sdk_vendor\n")
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(policy_file=pf, policy_file_sha256="ab" * 32)
-        )
-        assert cfg.provenance[POLICY_BASE_FIELD].sha256 == "ab" * 32
-
-    def test_a_directly_constructed_policy_file_records_no_digest(self):
-        # No parse, no digest to claim -- the receipt says so rather than
-        # inventing one by reading whatever path it was handed.
-        cfg = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=PolicyFile()))
-        assert cfg.provenance[POLICY_BASE_FIELD].sha256 is None
-
-    def test_a_pack_filled_field_names_the_manifest_that_filled_it(self, tmp_path):
-        pack = _write_pack(
-            tmp_path / "c.yml",
-            pack_id="rust_c_ffi",
-            kind="contract",
-            assignments="contract.unresolved: warn\n",
-        )
-        cfg = _resolve(explicit=ExplicitCompatibilityInputs(pack_paths=(str(pack),)))
-        prov = cfg.provenance["contract.unresolved"]
-        assert prov.reference == "rust_c_ffi"
-        assert prov.version == 1
-        assert prov.path == str(pack)
-        assert prov.sha256 == hashlib.sha256(pack.read_bytes()).hexdigest()
-        assert prov.field_location == "contract.unresolved"
-
-    def test_the_recorded_pack_does_not_depend_on_path_order(self, tmp_path):
-        # Two packs agreeing on a field are equals; which one the receipt
-        # names must still not depend on the order they were passed in.
-        first = _write_pack(
-            tmp_path / "a.yml",
-            pack_id="a",
-            kind="contract",
-            assignments="contract.unresolved: warn\n",
-        )
-        second = _write_pack(
-            tmp_path / "b.yml",
-            pack_id="b",
-            kind="contract",
-            assignments="contract.unresolved: warn\n",
-        )
-        forward = _resolve(
-            explicit=ExplicitCompatibilityInputs(pack_paths=(str(first), str(second)))
-        )
-        reverse = _resolve(
-            explicit=ExplicitCompatibilityInputs(pack_paths=(str(second), str(first)))
-        )
-        assert forward == reverse
-        assert forward.provenance["contract.unresolved"].reference == "a"
-
-
-class TestOverridesContributors:
-    """`policy.overrides` is composed; the receipt lists real contributors only."""
-
-    def test_a_non_policy_pack_is_not_credited(self, tmp_path):
-        gate_pack = _write_pack(
-            tmp_path / "g.yml",
-            pack_id="g",
-            kind="gate",
-            assignments="gate.exit_code_scheme: severity\n",
-        )
-        pf = _policy_file(
-            tmp_path,
-            """\
-            overrides:
-              soname_changed: warn
-            """,
-        )
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                policy_file=pf, pack_paths=(str(gate_pack),)
-            )
-        )
-        options = {e.option for e in cfg.provenance[POLICY_OVERRIDES_FIELD].selected_by}
-        assert options == {"--policy-file"}
-
-    def test_a_fully_shadowed_policy_pack_is_not_credited(self, tmp_path):
-        pack = _write_pack(
-            tmp_path / "p.yml",
-            pack_id="p",
-            kind="policy",
-            assignments="soname_changed: break\n",
-        )
-        pf = _policy_file(
-            tmp_path,
-            """\
-            overrides:
-              soname_changed: ignore
-            """,
-        )
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                policy_file=pf, pack_paths=(str(pack),)
-            )
-        )
-        assert cfg.policy.overrides["soname_changed"] is Verdict.COMPATIBLE
-        options = {e.option for e in cfg.provenance[POLICY_OVERRIDES_FIELD].selected_by}
-        assert options == {"--policy-file"}
-
-    def test_a_contributing_policy_pack_is_credited(self, tmp_path):
-        pack = _write_pack(
-            tmp_path / "p.yml",
-            pack_id="p",
-            kind="policy",
-            assignments="func_added: warn\n",
-        )
-        pf = _policy_file(
-            tmp_path,
-            """\
-            overrides:
-              soname_changed: ignore
-            """,
-        )
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                policy_file=pf, pack_paths=(str(pack),)
-            )
-        )
-        options = {e.option for e in cfg.provenance[POLICY_OVERRIDES_FIELD].selected_by}
-        assert options == {"--policy-file", "--pack"}
-
-
-class TestPathInputsAreNotSilentlyIgnored:
-    """A path an invocation named must reach the resolved configuration."""
-
-    def test_public_symbols_list_alone_resolves_a_real_explicit_scope(self, tmp_path):
-        listed = tmp_path / "symbols.txt"
-        listed.write_text("# comment\n\nfrom_file\n")
-        cfg = _resolve(
-            explicit=compare_cli_inputs({"public_symbols_list": listed})
-        )
-        assert cfg.surface.explicit_scope is not None
-        assert cfg.surface.explicit_scope.items == ("from_file",)
-
-    def test_inline_symbols_and_the_list_file_merge(self, tmp_path):
-        listed = tmp_path / "symbols.txt"
-        listed.write_text("from_file\n")
-        cfg = _resolve(
-            explicit=compare_cli_inputs(
-                {"public_symbols": ("inline",), "public_symbols_list": listed}
-            )
-        )
-        assert cfg.surface.explicit_scope is not None
-        assert cfg.surface.explicit_scope.items == ("from_file", "inline")
-
-    def test_the_two_symbol_sources_stay_separately_attributed(self, tmp_path):
-        # Flattening them lost which symbols came from the file, leaving a
-        # receipt that named only --public-symbol and no path.
-        listed = tmp_path / "symbols.txt"
-        listed.write_text("from_file\n")
-        inputs = compare_cli_inputs(
-            {"public_symbols": ("inline",), "public_symbols_list": listed}
-        )
-        assert inputs.public_symbols == ("inline",)
-        assert inputs.public_symbols_list is not None
-        assert inputs.public_symbols_list.items == ("from_file",)
-
-        cfg = _resolve(explicit=inputs)
-        entries = {
-            (e.option, e.path) for e in cfg.provenance[EXPLICIT_SCOPE_FIELD].selected_by
-        }
-        assert entries == {
-            ("--public-symbol", None),
-            ("--public-symbols-list", str(listed)),
-        }
-
-    def test_a_list_only_invocation_names_the_file_in_its_receipt(self, tmp_path):
-        listed = tmp_path / "symbols.txt"
-        listed.write_text("from_file\n")
-        cfg = _resolve(explicit=compare_cli_inputs({"public_symbols_list": listed}))
-        selected_by = cfg.provenance[EXPLICIT_SCOPE_FIELD].selected_by
-        assert [(e.option, e.path) for e in selected_by] == [
-            ("--public-symbols-list", str(listed))
-        ]
-
-    def test_cli_policy_file_path_is_loaded_when_not_pre_supplied(self, tmp_path):
-        path = tmp_path / "policy.yml"
-        path.write_text("base_policy: sdk_vendor\n")
-        cfg = _resolve(explicit=compare_cli_inputs({"policy_file_path": path}))
-        assert cfg.policy.base.id == "sdk_vendor"
-
-    def test_a_pre_supplied_policy_file_is_not_re_read(self, tmp_path):
-        pf = _policy_file(tmp_path, "base_policy: plugin_abi\n")
-        cfg = _resolve(
-            explicit=compare_cli_inputs(
-                {"policy_file_path": tmp_path / "other.yml"}, policy_file=pf
-            )
-        )
-        assert cfg.policy.base.id == "plugin_abi"
-
-    def test_cli_suppress_path_is_read_when_not_pre_supplied(self, tmp_path):
-        path = tmp_path / "s.yml"
-        path.write_text("version: 1\nsuppressions:\n  - symbol: foo\n    reason: ok\n")
-        cfg = _resolve(explicit=compare_cli_inputs({"suppress": path}))
-        assert cfg.suppressions is not None
-        assert cfg.suppressions.rules == ("symbol='foo'",)
-
-    def test_request_policy_and_suppression_paths_are_honoured(self, tmp_path):
-        policy = tmp_path / "policy.yml"
-        policy.write_text("base_policy: sdk_vendor\n")
-        suppress = tmp_path / "s.yml"
-        suppress.write_text(
-            "version: 1\nsuppressions:\n  - symbol: foo\n    reason: ok\n"
-        )
-        request = CompareRequest(
-            old=InputSpec(path=tmp_path / "old.so"),
-            new=InputSpec(path=tmp_path / "new.so"),
-            policy_file_path=policy,
-            suppress=suppress,
-        )
-        cfg = compatibility_config_from_compare_request(request)
-        # Previously both request fields were ignored unless the caller
-        # separately re-loaded the same files.
-        assert cfg.policy.base.id == "sdk_vendor"
-        assert cfg.suppressions is not None
-        assert cfg.suppressions.rules == ("symbol='foo'",)
-
-
-class TestReceiptMetadataIsCompared:
-    def test_a_differing_source_digest_is_a_difference(self, tmp_path):
-        pf = _policy_file(tmp_path, "base_policy: sdk_vendor\n")
-        one = _resolve(explicit=ExplicitCompatibilityInputs(policy_file=pf))
-        two = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                policy_file=pf, policy_file_sha256="cd" * 32
-            )
-        )
-        # Same resolved base, different recorded digest: not equivalent.
-        assert one.policy == two.policy
-        differences = cross_front_end_differences(one, two)
-        assert any(POLICY_BASE_FIELD in d for d in differences)
-
-    def test_a_differing_selected_by_path_is_a_difference(self, tmp_path):
-        first = _write_pack(
-            tmp_path / "a.yml",
-            pack_id="p",
-            kind="contract",
-            assignments="contract.unresolved: warn\n",
-        )
-        copied = tmp_path / "copy.yml"
-        copied.write_text(first.read_text())
-        a = _resolve(explicit=ExplicitCompatibilityInputs(pack_paths=(str(first),)))
-        b = _resolve(explicit=ExplicitCompatibilityInputs(pack_paths=(str(copied),)))
-        assert a.contract == b.contract
-        assert cross_front_end_differences(a, b)
-
-    def test_the_option_spelling_alone_is_still_permitted_to_differ(self, tmp_path):
-        # The genuine cross-front-end case the normalization exists for.
-        policy = tmp_path / "policy.yml"
-        policy.write_text("base_policy: sdk_vendor\n")
-        cli = _resolve(
-            front_end=FrontEnd.CLI,
-            explicit=compare_cli_inputs(
-                {
-                    "policy_file_path": policy,
-                    "policy": "strict_abi",
-                    # A typed request has no "unset" for scope_public, so the
-                    # equivalent CLI run has to state the flag too.
-                    "scope_public_headers": True,
-                },
-                explicit_parameters={"policy", "scope_public_headers"},
-            ),
-        )
-        api = compatibility_config_from_compare_request(
-            CompareRequest(
-                old=InputSpec(path=tmp_path / "old.so"),
-                new=InputSpec(path=tmp_path / "new.so"),
-                policy_file_path=policy,
-                policy="strict_abi",
-            )
-        )
-        assert cross_front_end_differences(cli, api) == []
-
-
-class TestProjectConfigProvenanceNamesItsSource:
-    def test_contract_mode_from_a_project_config_names_the_key_and_file(self, tmp_path):
-        # A PROJECT_CONFIG receipt claiming `--scope-public-headers`, with no
-        # path, cannot identify what a replay would have to re-read.
-        cfg = _resolve(
-            project=ProjectCompatibilityInputs(
-                path=str(tmp_path / ".abicheck.yml"), scope_public=False
-            )
-        )
-        prov = cfg.provenance[CONTRACT_MODE_FIELD]
-        assert prov.layer is SelectorLayer.PROJECT_CONFIG
-        assert prov.path == str(tmp_path / ".abicheck.yml")
-        assert [(e.option, e.path) for e in prov.selected_by] == [
-            ("scope.public", str(tmp_path / ".abicheck.yml"))
-        ]
-
-    def test_the_cli_flag_still_records_its_own_spelling(self):
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(scope_public_headers=False)
-        )
-        prov = cfg.provenance[CONTRACT_MODE_FIELD]
-        assert prov.layer is SelectorLayer.LEGACY_ALIAS
-        assert prov.path is None
-        assert [e.option for e in prov.selected_by] == ["--no-scope-public-headers"]
-
-
-class TestManifestsAreReadOncePerResolution:
-    def test_each_manifest_is_loaded_exactly_once(self, tmp_path, monkeypatch):
-        # Re-reading per resolver call is wasteful, and a manifest edited
-        # between those reads would leave one configuration's receipt entries
-        # carrying different identities for the same pack.
-        import abicheck.compatibility_evaluation_wiring as wiring
-
-        pack = _write_pack(
-            tmp_path / "c.yml",
-            pack_id="p",
-            kind="contract",
-            assignments="contract.unresolved: warn\n",
-        )
-        reads: list[str] = []
-        real = wiring.load_pack_manifest
-
-        def _counting(path):
-            reads.append(str(path))
-            return real(path)
-
-        monkeypatch.setattr(wiring, "load_pack_manifest", _counting)
-        cfg = _resolve(explicit=ExplicitCompatibilityInputs(pack_paths=(str(pack),)))
-        assert cfg.contract.unresolved == "warn"
-        assert reads == [str(pack)]
-
-    def test_the_suppression_receipt_carries_its_source_digest(self, tmp_path):
-        path = tmp_path / "s.yml"
-        path.write_text("version: 1\nsuppressions:\n  - symbol: foo\n    reason: ok\n")
-        source = SuppressionSource.from_file(path)
-        cfg = _resolve(explicit=ExplicitCompatibilityInputs(suppression=source))
-        assert cfg.provenance[SUPPRESSIONS_FIELD].sha256 == source.sha256
-
-
 class TestSymbolListParsing:
     def test_the_shared_merge_helper_reads_both_sources(self, tmp_path):
         # `cli_helpers_compare` re-exports this as `_collect_force_public_symbols`,
@@ -1368,118 +962,3 @@ class TestSymbolListParsing:
 
     def test_no_list_file_means_inline_symbols_only(self):
         assert collect_force_public_symbols(("a", "b"), None) == {"a", "b"}
-
-
-class TestOverridesReceiptNamesEachContributingPack:
-    def _pack(self, tmp_path, name, assignments):
-        return _write_pack(
-            tmp_path / f"{name}.yml",
-            pack_id=name,
-            kind="policy",
-            assignments=assignments,
-        )
-
-    def test_each_contributing_pack_gets_its_own_entry_with_its_path(self, tmp_path):
-        first = self._pack(tmp_path, "a", "soname_changed: break\n")
-        second = self._pack(tmp_path, "b", "func_added: warn\n")
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                pack_paths=(str(first), str(second))
-            )
-        )
-        entries = [
-            (e.option, e.path)
-            for e in cfg.provenance[POLICY_OVERRIDES_FIELD].selected_by
-        ]
-        assert entries == [("--pack", str(first)), ("--pack", str(second))]
-
-    def test_a_fully_shadowed_pack_is_not_listed(self, tmp_path):
-        shadowed = self._pack(tmp_path, "a", "soname_changed: break\n")
-        contributing = self._pack(tmp_path, "b", "func_added: warn\n")
-        pf = _policy_file(
-            tmp_path,
-            """\
-            overrides:
-              soname_changed: ignore
-            """,
-        )
-        cfg = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                policy_file=pf, pack_paths=(str(shadowed), str(contributing))
-            )
-        )
-        entries = [
-            (e.option, e.path)
-            for e in cfg.provenance[POLICY_OVERRIDES_FIELD].selected_by
-        ]
-        assert entries == [
-            ("--policy-file", str(tmp_path / "policy.yml")),
-            ("--pack", str(contributing)),
-        ]
-
-    def test_a_lone_pack_source_is_named_outright(self, tmp_path):
-        only = self._pack(tmp_path, "a", "func_added: warn\n")
-        prov = _resolve(
-            explicit=ExplicitCompatibilityInputs(pack_paths=(str(only),))
-        ).provenance[POLICY_OVERRIDES_FIELD]
-        assert prov.reference == "a"
-        assert prov.version == 1
-        assert prov.path == str(only)
-        assert prov.sha256 == hashlib.sha256(only.read_bytes()).hexdigest()
-
-    def test_several_pack_sources_claim_no_single_manifest(self, tmp_path):
-        first = self._pack(tmp_path, "a", "soname_changed: break\n")
-        second = self._pack(tmp_path, "b", "func_added: warn\n")
-        prov = _resolve(
-            explicit=ExplicitCompatibilityInputs(
-                pack_paths=(str(first), str(second))
-            )
-        ).provenance[POLICY_OVERRIDES_FIELD]
-        # No one manifest supplied the whole value, so none is named as *the*
-        # source -- every contributor is in selected_by instead.
-        assert prov.path is None
-        assert prov.reference is None
-        assert len(prov.selected_by) == 2
-
-
-class TestSelectedButEmptySources:
-    def test_an_empty_symbols_list_is_still_a_selected_source(self, tmp_path):
-        # "A source was selected and resolved to zero items" is a different
-        # fact from "no source was selected" -- and the path must survive.
-        listed = tmp_path / "symbols.txt"
-        listed.write_text("# only a comment\n\n")
-        cfg = _resolve(explicit=compare_cli_inputs({"public_symbols_list": listed}))
-        assert cfg.surface.explicit_scope is not None
-        assert cfg.surface.explicit_scope.items == ()
-        prov = cfg.provenance[EXPLICIT_SCOPE_FIELD]
-        assert prov.layer is SelectorLayer.EXPLICIT_CLI
-        assert [(e.option, e.path) for e in prov.selected_by] == [
-            ("--public-symbols-list", str(listed))
-        ]
-
-    def test_no_source_at_all_stays_none(self):
-        cfg = _resolve()
-        assert cfg.surface.explicit_scope is None
-        assert (
-            cfg.provenance[EXPLICIT_SCOPE_FIELD].layer
-            is SelectorLayer.BUILT_IN_DEFAULT
-        )
-
-
-class TestPolicyDigestIsOverRawBytes:
-    def test_line_ending_only_changes_are_detected(self, tmp_path):
-        # `read_text()` translates newlines, so hashing its result would give
-        # a CRLF file and its LF twin the same digest.
-        path = tmp_path / "policy.yml"
-        path.write_bytes(b"base_policy: sdk_vendor\r\n")
-        crlf = PolicyFile.load(path)
-        path.write_bytes(b"base_policy: sdk_vendor\n")
-        lf = PolicyFile.load(path)
-        assert crlf.base_policy == lf.base_policy == "sdk_vendor"
-        assert crlf.source_sha256 != lf.source_sha256
-
-    def test_the_digest_matches_the_files_raw_bytes(self, tmp_path):
-        path = tmp_path / "policy.yml"
-        path.write_bytes(b"base_policy: plugin_abi\r\n")
-        loaded = PolicyFile.load(path)
-        assert loaded.source_sha256 == hashlib.sha256(path.read_bytes()).hexdigest()
